// src/services/clever-framework/biz-typings.d.ts
// @ts-ignore
/* eslint-disable */

import { Protocol } from 'puppeteer-core';

declare namespace Biz {
  type News = {
    title?: string;
    showNum?: string;
    likeNum?: number;
    showStatus?: string;
    showTime?: string;
    localDate?: number;
    localTime?: string;
    localDateTime?: string;
  }

}